requirements.txt
sanic
image
pillow-simd
redis
nvgpu
opencv-python
tensorflow-gpu
pybase64
